requirements.txt
langchain==0.1.8
python-dotenv==1.0.0
huggingface_hub
typing

# pdf
PyPDF2==3.0.1
pypdf==3.17.3

# embedding
InstructorEmbedding==1.0.1
torch==2.2.2
tqdm==4.66.2
sentence-transformers==2.2.2
faiss-cpu

# mongodb
pymongo==4.6.1

cloudinary

# API-END point
fastapi==0.109.2
fastapi-cors
uvicorn[standard]==0.17.*
python-multipart==0.0.9
PyJWT==2.8.0
pydantic



# requests-toolbelt